move tests that expect a legitimate GitHub token to test-api changing these test to use mocks will take more work and may invalidate them there is still some interaction with GitHub in the unit tests but any string without whitespace can be used as a token
