+++ test-api/gitHubTest.ts
@@ -1,9 +1,16 @@
 import "mocha";
 
 import * as assert from "power-assert";
-import { GitHubRepoRef } from "../../../src/operations/common/GitHubRepoRef";
-import { createCommitComment, deepLink, fileContent, hasFile } from "../../../src/util/gitHub";
-import { GitHubToken } from "../../credentials";
+import { GitHubRepoRef } from "../src/operations/common/GitHubRepoRef";
+import { createCommitComment, deepLink, fileContent, hasFile } from "../src/util/gitHub";
+
+function barf(): string {
+    throw new Error("<please set GITHUB_TOKEN env variable>");
+}
+
+export const GitHubToken: string = process.env.GITHUB_TOKEN || barf();
+
+export const Creds = { token: GitHubToken };
 
 describe("gitHubUtils", () => {
