+++ test/project/git/GitProjectTest.ts
@@ -117,46 +117,6 @@ describe("GitProject", () => {
             .then(() => done(), done);
     });
 
-    it("add a file, init and commit, then push to new remote repo", function(done) {
-        this.retries(5);
-
-        const p = tempProject();
-        p.addFileSync("Thing", "1");
-
-        const repo = `test-repo-2-${new Date().getTime()}`;
-
-        const gp: GitProject = GitCommandGitProject.fromProject(p, Creds);
-
-        getOwnerByToken().then(owner => gp.init()
-            .then(() => gp.createAndSetGitHubRemote(owner, repo, "Thing1", TestRepositoryVisibility))
-            .then(() => gp.commit("Added a Thing"))
-            .then(() => gp.push()
-                .then(() => deleteRepoIfExists({ owner, repo }).then(done)),
-        ).catch(() => deleteRepoIfExists({ owner, repo }).then(done)),
-        ).catch(done);
-
-    }).timeout(16000);
-
-    it("add a file, then PR push to remote repo", function(done) {
-        this.retries(1);
-
-        newRepo()
-            .then(ownerAndRepo => GitCommandGitProject.cloned(Creds,
-                new GitHubRepoRef(ownerAndRepo.owner, ownerAndRepo.repo))
-                .then(gp => {
-                    gp.addFileSync("Cat", "hat");
-                    const branch = "thing2";
-                    return gp.createBranch(branch)
-                        .then(() => gp.commit("Added a Thing"))
-                        .then(() => gp.push())
-                        .then(() => gp.raisePullRequest("Thing2", "Adds another character"))
-                        .then(() => deleteRepoIfExists(ownerAndRepo));
-                }).catch(err => deleteRepoIfExists(ownerAndRepo)
-                    .then(() => Promise.reject(err))))
-            .then(() => done(), done);
-
-    }).timeout(20000);
-
     it("check out commit", done => {
         const sha = "590ed8f7a2430d45127ea04cc5bdf736fe698712";
         GitCommandGitProject.cloned(Creds, new GitHubRepoRef("atomist", "microgrammar", sha))
@@ -183,61 +143,3 @@ describe("GitProject", () => {
     }).timeout(10000);
 
 });
-
-/**
- * Create a new repo we can use for tests
- * @return {Promise<{owner: string; repo: string}>}
- */
-export function newRepo(): Promise<{ owner: string, repo: string }> {
-    const config = {
-        headers: {
-            Authorization: `token ${GitHubToken}`,
-        },
-    };
-    const name = `test-repo-${new Date().getTime()}`;
-    const description = "a thing";
-    const url = `${GitHubDotComBase}/user/repos`;
-    console.debug("Visibility is " + TestRepositoryVisibility);
-    return getOwnerByToken()
-        .then(owner => axios.post(url, {
-            name,
-            description,
-            private: TestRepositoryVisibility === "private",
-            auto_init: true,
-        }, config)
-            .then(() =>
-                ({ owner, repo: name })))
-        .catch(error => {
-            if (error.response.status === 422) {
-                throw new Error("Could not create repository. GitHub says: " +
-                    _.get(error, "response.data.message", "nothing"));
-            } else {
-                throw new Error("Could not create repo: " + error.message);
-            }
-        });
-}
-
-export function deleteRepoIfExists(ownerAndRepo: { owner: string, repo: string }): Promise<any> {
-    console.debug("Cleanup: deleting " + ownerAndRepo.repo);
-    const config = {
-        headers: {
-            Authorization: `token ${GitHubToken}`,
-        },
-    };
-    const url = `${GitHubDotComBase}/repos/${ownerAndRepo.owner}/${ownerAndRepo.repo}`;
-    return axios.delete(url, config)
-        .catch(err => {
-            console.error(`error deleting ${ownerAndRepo.repo}, ignoring. ${err.response.status}`);
-        });
-}
-
-function getOwnerByToken(): Promise<string> {
-    const config = {
-        headers: {
-            Authorization: `token ${GitHubToken}`,
-        },
-    };
-    return axios.get(`${GitHubDotComBase}/user`, config).then(response =>
-        response.data.login,
-    );
-}
